requirements.txt
fastapi
sqlalchemy
uvicorn
python-dotenv
pydantic